основное задание по датам
