+++ init.php
@@ -1,0 +1,7 @@
+<?php
+$is_auth = rand(0, 1);
+$user_name = 'Anna';
+date_default_timezone_set('Europe/Moscow');
+
+// Константы
+define('HOURS_PER_DAY', 24);
